tn discrepancies endpoint
