+++ discovery-provider/src/queries/health_check.py
@@ -9,6 +9,7 @@ from src.queries.get_db_seed_restore_status import get_db_seed_restore_status
 from src.queries.get_health import get_health, get_location
 from src.queries.get_latest_play import get_latest_play
 from src.queries.get_sol_plays import get_latest_sol_play_check_info
+from src.queries.get_trusted_notifier_discrepancies import get_trusted_notifier_discrepancies
 from src.queries.queries import parse_bool_param
 from src.tasks.index_profile_challenge_backfill import (
     index_profile_challenge_backfill_tablename,
@@ -61,6 +62,12 @@ def health_check():
     }
 
     (health_results, error) = get_health(args)
+    return success_response(health_results, 500 if error else 200, sign_response=False)
+
+
+@bp.route("/trusted_notifier_discrepancies_check", methods=["GET"])
+def trusted_notifier_discrepancies_check():
+    (health_results, error) = get_trusted_notifier_discrepancies()
     return success_response(health_results, 500 if error else 200, sign_response=False)
